refactor: extract Provider interface per provider
Moves Google, Replicate, and debug provider logic into dedicated
files under lib/providers/ behind a common Provider interface.
lib/generation.ts, lib/textModel.ts, and lib/upscaling.ts are now
thin dispatchers that look up the provider in a registry.

inferName/inferIcon move to lib/modelNames.ts so both providers can
share model-id normalization. lib/replicateSchema.ts is folded into
providers/replicate.ts.

promptVariations now reads supportsTextPrefill from the Provider
instead of hardcoding "provider === google".

settingsStore.addCustomModel takes a provider param so future
provider-owned custom models (e.g. Gemini image models added via
search) can be stored with the correct id prefix.

Behavior-preserving. No data migration needed.

Towards #34
