+++ app/lib/providers/google.ts
@@ -1,0 +1,160 @@
+import { GoogleGenAI, ThinkingLevel } from "@google/genai";
+import type { GenerationParams, GenerationResult } from "~/lib/generation";
+import { getImageDimensions } from "~/lib/imageProcessing";
+import { toRateLimitError } from "~/lib/retry";
+import { blobToBase64 } from "~/lib/util";
+import type { Provider, TextGenerationArgs } from "./types";
+
+async function generateImage(params: GenerationParams, apiKey?: string): Promise<GenerationResult> {
+  if (!apiKey) throw new Error("No API key for google");
+  const ai = new GoogleGenAI({ apiKey });
+
+  const parts: Array<{ text?: string; inlineData?: { mimeType: string; data: string } }> = [];
+
+  for (const ref of params.referenceImages) {
+    const base64 = await blobToBase64(ref.blob);
+    parts.push({
+      inlineData: {
+        mimeType: "image/png",
+        data: base64.split(",")[1],
+      },
+    });
+  }
+  parts.push({ text: params.prompt });
+
+  const config = {
+    responseModalities: ["IMAGE"],
+    ...(params.aspectRatio || params.resolution
+      ? {
+          imageConfig: {
+            ...(params.aspectRatio && { aspectRatio: params.aspectRatio }),
+            ...(params.resolution && { imageSize: params.resolution }),
+          },
+        }
+      : {}),
+  };
+
+  let response;
+  try {
+    response = await ai.models.generateContentStream({
+      model: params.modelId,
+      config,
+      contents: [{ role: "user", parts }],
+    });
+  } catch (error) {
+    throw toRateLimitError(error, "google");
+  }
+
+  let imageBlob: Blob | null = null;
+  let modelVersion: string | undefined;
+
+  for await (const chunk of response) {
+    if (!chunk.candidates?.[0]?.content?.parts) continue;
+    const inlineData = chunk.candidates[0].content.parts[0]?.inlineData;
+    if (inlineData?.data && inlineData?.mimeType) {
+      const binaryString = atob(inlineData.data);
+      const bytes = new Uint8Array(binaryString.length);
+      for (let i = 0; i < binaryString.length; i++) bytes[i] = binaryString.charCodeAt(i);
+      imageBlob = new Blob([bytes], { type: inlineData.mimeType });
+    }
+    if (chunk.modelVersion) modelVersion = chunk.modelVersion;
+  }
+
+  if (!imageBlob) throw new Error("No image in response");
+  const dimensions = await getImageDimensions(imageBlob);
+
+  return {
+    blob: imageBlob,
+    width: dimensions.width,
+    height: dimensions.height,
+    metadata: { modelVersion },
+  };
+}
+
+async function generateText(args: TextGenerationArgs, apiKey: string): Promise<string> {
+  const { modelId, systemPrompt, userPrompt, images, prefill } = args;
+  const ai = new GoogleGenAI({ apiKey });
+
+  const parts: Array<{ text?: string; inlineData?: { mimeType: string; data: string } }> = [];
+
+  if (images?.length) {
+    for (const blob of images) {
+      const base64 = await blobToBase64(blob);
+      parts.push({
+        inlineData: {
+          mimeType: blob.type || "image/png",
+          data: base64.split(",")[1],
+        },
+      });
+    }
+  }
+
+  parts.push({ text: userPrompt });
+
+  const contents: Array<{
+    role: "user" | "model";
+    parts: Array<{ text?: string; inlineData?: { mimeType: string; data: string } }>;
+  }> = [{ role: "user", parts }];
+
+  if (prefill) {
+    contents.push({ role: "model", parts: [{ text: prefill }] });
+  }
+
+  let response;
+  try {
+    response = await ai.models.generateContent({
+      model: modelId,
+      config: {
+        systemInstruction: systemPrompt,
+        thinkingConfig: {
+          thinkingLevel: ThinkingLevel.LOW,
+        },
+      },
+      contents,
+    });
+  } catch (error) {
+    throw toRateLimitError(error, "google");
+  }
+
+  const text = response.text;
+  if (!text) {
+    throw new Error("No text in response");
+  }
+
+  if (prefill) {
+    return prefill + text;
+  }
+
+  return text;
+}
+
+async function testTextModel(apiKey: string, modelId: string): Promise<void> {
+  await generateText(
+    {
+      modelId,
+      systemPrompt: "You are a connectivity test.",
+      userPrompt: "Respond with the single word 'hi' and nothing else.",
+    },
+    apiKey
+  );
+}
+
+export const googleProvider: Provider = {
+  id: "google",
+  label: "Google",
+  iconPath: "/icons/google.svg",
+  requiresApiKey: true,
+  supportsTextPrefill: true,
+  capabilities: {
+    image: true,
+    text: true,
+    upscale: false,
+    searchImage: false,
+    searchText: false,
+    searchUpscale: false,
+    resolveImageModel: false,
+  },
+  generateImage,
+  generateText,
+  testTextModel,
+};
