+++ app/lib/textModel.ts
@@ -1,10 +1,8 @@
-import { GoogleGenAI, ThinkingLevel } from "@google/genai";
-import Replicate from "replicate";
 import { useSettingsStore } from "~/stores/settingsStore";
 import type { ApiKeyProvider } from "~/types";
-import { blobToBase64 } from "./util";
+import { getProvider } from "~/lib/providers";
 import { logger } from "./logging";
-import { retryWithBackoff, toRateLimitError } from "./retry";
+import { retryWithBackoff } from "./retry";
 
 const DEFAULT_GOOGLE_MODEL = "gemini-3-flash-preview";
 const DEFAULT_REPLICATE_MODEL = "google/gemini-3-flash";
@@ -23,7 +21,6 @@ function resolveProvider(): ResolvedProvider {
     throw new Error("No text model configured");
   }
 
-  // Try selected model's provider first
   if (apiKeys[selected.provider]) {
     return {
       provider: selected.provider,
@@ -32,7 +29,6 @@ function resolveProvider(): ResolvedProvider {
     };
   }
 
-  // Fallback to the other provider with its default model
   const fallback: ApiKeyProvider = selected.provider === "google" ? "replicate" : "google";
   if (apiKeys[fallback]) {
     return {
@@ -53,7 +49,7 @@ export function isTextModelAvailable(): boolean {
 /**
  * Returns the provider that would be used for the next `callTextModel` call,
  * or null if no API keys are available. Useful for call sites that want to
- * branch on capability (e.g. only use response prefill on the Google path).
+ * branch on capability.
  */
 export function resolveTextModelProvider(): ApiKeyProvider | null {
   try {
@@ -77,121 +73,17 @@ export async function callTextModel(
     `\nPrefill: ${prefill ?? "none"}`
   );
 
-  const output = await retryWithBackoff(() => {
-    let output = "";
-    if (provider === "google") {
-      return callGoogleTextModel(apiKey, modelId, systemPrompt, userPrompt, images, prefill);
-    }
-    return callReplicateTextModel(apiKey, modelId, systemPrompt, userPrompt, images, prefill);
-  });
+  const impl = getProvider(provider).generateText;
+  if (!impl) {
+    throw new Error(`Provider ${provider} does not support text generation`);
+  }
+
+  const output = await retryWithBackoff(() =>
+    impl({ modelId, systemPrompt, userPrompt, images, prefill }, apiKey)
+  );
 
   logger.debug("Text model output:", output);
   return output;
-}
-
-async function callGoogleTextModel(
-  apiKey: string,
-  modelId: string,
-  systemPrompt: string,
-  userPrompt: string,
-  images?: Blob[],
-  prefill?: string
-): Promise<string> {
-  const ai = new GoogleGenAI({ apiKey });
-
-  const parts: Array<{ text?: string; inlineData?: { mimeType: string; data: string } }> = [];
-
-  if (images?.length) {
-    for (const blob of images) {
-      const base64 = await blobToBase64(blob);
-      parts.push({
-        inlineData: {
-          mimeType: blob.type || "image/png",
-          data: base64.split(",")[1],
-        },
-      });
-    }
-  }
-
-  parts.push({ text: userPrompt });
-
-  const contents: Array<{
-    role: "user" | "model";
-    parts: Array<{ text?: string; inlineData?: { mimeType: string; data: string } }>;
-  }> = [{ role: "user", parts }];
-
-  if (prefill) {
-    contents.push({ role: "model", parts: [{ text: prefill }] });
-  }
-
-  let response;
-  try {
-    response = await ai.models.generateContent({
-      model: modelId,
-      config: {
-        systemInstruction: systemPrompt,
-        thinkingConfig: {
-          thinkingLevel: ThinkingLevel.LOW,
-        },
-      },
-      contents,
-    });
-  } catch (error) {
-    throw toRateLimitError(error, "google");
-  }
-
-  const text = response.text;
-  if (!text) {
-    throw new Error("No text in response");
-  }
-
-  if (prefill) {
-    logger.debug(
-      "Google text model response with prefill. prefill=",
-      prefill,
-      "response.text=",
-      text
-    );
-    return prefill + text;
-  }
-
-  return text;
-}
-
-async function callReplicateTextModel(
-  apiKey: string,
-  modelId: string,
-  systemPrompt: string,
-  userPrompt: string,
-  images?: Blob[],
-  prefill?: string
-): Promise<string> {
-  const baseUrl = new URL("/proxy/replicate/v1", window.location.origin).toString();
-  const replicate = new Replicate({ auth: apiKey, baseUrl });
-
-  // Replicate doesn't support prefills, so we just concatenate it to the user prompt. The model should be able to handle this as long as the prefill is clearly separated from the user prompt (e.g. with a newline). true solution is to use resolveTextProvier() and not send prefill to replicate at all.
-  if (prefill) {
-    userPrompt = prefill + "\n" + userPrompt;
-  }
-
-  let output;
-  try {
-    output = await replicate.run(modelId as `${string}/${string}`, {
-      input: { prompt: userPrompt, system_instruction: systemPrompt, images },
-    });
-  } catch (error) {
-    throw toRateLimitError(error, "replicate");
-  }
-
-  if (typeof output === "string") {
-    return output;
-  }
-
-  if (Array.isArray(output)) {
-    return output.join("");
-  }
-
-  throw new Error("Unexpected text model response format");
 }
 
 /**
@@ -204,11 +96,9 @@ export async function testTextModel(
   apiKey: string,
   modelId: string
 ): Promise<void> {
-  const systemPrompt = "You are a connectivity test.";
-  const userPrompt = "Respond with the single word 'hi' and nothing else.";
-  if (provider === "google") {
-    await callGoogleTextModel(apiKey, modelId, systemPrompt, userPrompt);
-  } else {
-    await callReplicateTextModel(apiKey, modelId, systemPrompt, userPrompt);
+  const impl = getProvider(provider).testTextModel;
+  if (!impl) {
+    throw new Error(`Provider ${provider} does not support text generation`);
   }
+  await impl(apiKey, modelId);
 }
