+++ app/lib/upscaling.ts
@@ -1,6 +1,4 @@
-import Replicate from "replicate";
-import { blobToBase64 } from "./util";
-import { getImageDimensions } from "./imageProcessing";
+import { getProvider } from "~/lib/providers";
 import type { GenerationResult } from "./generation";
 import type { StoredUpscaler } from "~/types";
 
@@ -9,33 +7,12 @@ export async function executeUpscale(
   upscaler: StoredUpscaler,
   apiKey: string
 ): Promise<GenerationResult> {
-  const baseUrl = new URL("/proxy/replicate/v1", window.location.origin).toString();
-  const replicate = new Replicate({ auth: apiKey, baseUrl });
-
-  const dataUri = await blobToBase64(sourceBlob);
-
-  const input: Record<string, unknown> = { image: dataUri };
-  if (upscaler.scaleParam !== null && upscaler.scale !== null) {
-    input[upscaler.scaleParam] = upscaler.scale;
+  // Upscalers are Replicate-only today; keep the lookup indirect so adding
+  // another provider's upscaler later is just a matter of extending the upscaler
+  // record with a provider field.
+  const provider = getProvider("replicate");
+  if (!provider.upscale) {
+    throw new Error("Replicate provider does not support upscaling");
   }
-
-  const output = await replicate.run(upscaler.replicateId as `${string}/${string}`, { input });
-
-  const imageUrl =
-    typeof output === "object" && output !== null && "url" in output
-      ? (output as { url: () => string }).url()
-      : Array.isArray(output)
-        ? output[0]
-        : String(output);
-
-  if (!imageUrl) throw new Error("No image in upscale response");
-
-  const imageResponse = await fetch(imageUrl);
-  if (!imageResponse.ok)
-    throw new Error(`Failed to fetch upscaled image: ${imageResponse.status}`);
-
-  const blob = await imageResponse.blob();
-  const dimensions = await getImageDimensions(blob);
-
-  return { blob, width: dimensions.width, height: dimensions.height, metadata: {} };
+  return provider.upscale(sourceBlob, upscaler, apiKey);
 }
